% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.R
\name{addSofa}
\alias{addSofa}
\title{Add SOFA scores.}
\usage{
addSofa(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{data.frame}}

\item{na.rm}{\code{logical}, should missing values replaced by zero?}
}
\value{
\code{data.frame}
}
\description{
Add SOFA scores to \code{data.frame}
}
