% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportSofa}
\alias{exportSofa}
\title{Write SOFA data.frame}
\usage{
exportSofa(x, path)
}
\arguments{
\item{x}{\code{data.frame}}

\item{path}{\code{character}, file path}
}
\description{
Write SOFA \code{data.frame} to file
}
