% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{icmsofa-package}
\alias{icmsofa}
\alias{icmsofa-package}
\title{icmsofa: Sequential Organ Failure Assessment (SOFA) Score}
\description{
This package is used to calculate Sequential Organ Failure
    Assessment (SOFA) Score from exported data of Dräger's ICM database.
}
\references{
\url{https://github.com/sgibb/icmsofa/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sgibb/icmsofa/}
  \item Report bugs at \url{https://github.com/sgibb/icmsofa/issues/}
}

}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de},
}
\keyword{package}
