% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importIcm}
\alias{importIcm}
\title{Import ICM data}
\usage{
importIcm(file, columns = c(CaseId = "HIS_CASEID", Date = "ADMINDATE",
  TreatmentId = "TREATMENTID", Description = "TREATMENTNAME", Type =
  "TREATMENTTYPE", Value = "NUMVALUE", Dose = "DOSE", Begin = "BEGIN", End
  = "END"), sep = "\\t", dec = ",", verbose = interactive())
}
\arguments{
\item{file}{\code{character}, filename}

\item{columns}{\code{character}, column names}

\item{sep}{\code{character}, field separator}

\item{dec}{\code{character}, decimal sign}

\item{verbose}{\code{logical}, verbose output?}
}
\value{
data.frame
}
\description{
Import ICM SQL-Export.
}
