% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importTimepoints}
\alias{importTimepoints}
\title{Import timepoint data.}
\usage{
importTimepoints(file, sep = "\\t", dec = ",", nonTimeColumns = 1L,
  format = "\%d.\%m.\%Y \%H:\%M")
}
\arguments{
\item{file}{\code{character}, filename}

\item{sep}{\code{character}, field separator}

\item{dec}{\code{character}, decimal sign}

\item{nonTimeColumns}{\code{numeric}, columns without time information}

\item{format}{\code{character}, time format}
}
\value{
data.frame
}
\description{
Import timepoint data.
}
