% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importXl}
\alias{importXl}
\title{Import ICM data}
\usage{
importXl(file, sheets = c(PAO2 = "PO2", FIO2 = "FIO", NOR = "Noradrenalin",
  DOB = "Dobutamin", IBP = "IBPm", BILI = "Bilirubin", PLT = "Thrombozyten",
  CREA = "Kreatinin"), columns = c(CaseId = "fallnummer", Date = "admindate",
  Description = "treatmentname", Value = "num"), verbose = interactive())
}
\arguments{
\item{file}{\code{character}, filename}

\item{sheets}{\code{character}, sheet names}

\item{columns}{\code{character}, column names}

\item{verbose}{\code{logical}, verbose output?}
}
\value{
data.frame
}
\description{
Import ICM XLS-Export.
}
