% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSofa}
\alias{plotSofa}
\title{Plot SOFA scores}
\usage{
plotSofa(x, path, timepoints)
}
\arguments{
\item{x}{\code{data.frame}}

\item{path}{\code{character}, file path}

\item{timepoints}{\code{data.frame}, with timepoints}
}
\description{
Plotting scores
}
