% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{sofa-package}
\alias{sofa}
\alias{sofa-package}
\title{sofa: Sequential Organ Failure Assessment (SOFA) Score}
\description{
Sequential Organ Failure Assessment (SOFA) Score
}
\references{
\url{https://github.com/sgibb/sofa/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sgibb/sofa/}
  \item Report bugs at \url{https://github.com/sgibb/sofa/issues/}
}

}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de},
}
\keyword{package}
