% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{sofaForTimepoints}
\alias{sofaForTimepoints}
\title{Get SOFA for specific timepoint}
\usage{
sofaForTimepoints(icm, tp)
}
\arguments{
\item{icm}{\code{data.frame}, ICM data}

\item{tp}{\code{data.frme}, timepoint data, first column Id, other columns
timepoints}
}
\value{
\code{data.frame}
}
\description{
Get SOFA for specific timepoint
}
