% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importIcm}
\alias{importIcm}
\title{Import ICM data}
\usage{
importIcm(file, sheets = c(PAO = "PO2", FIO = "FIO", NOR = "Noradrenalin", DOB
  = "Dobutamin", IBP = "IBPm", BIL = "Bilirubin", PLT = "Thrombozyten", CRE =
  "Kreatinin"), columns = c(Id = "fallnummer", Date = "admindate", Description
  = "treatmentname", Value = "num"))
}
\arguments{
\item{file}{\code{character}, filename}

\item{sheets}{\code{character}, sheet names}

\item{columns}{\code{character}, colum names}
}
\value{
data.frame
}
\description{
Import ICM SQL-Export.
}
