% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importTimepoints}
\alias{importTimepoints}
\title{Import timepoint data.}
\usage{
importTimepoints(file, sep = "\\t", dec = ",")
}
\arguments{
\item{file}{\code{character}, filename}

\item{sep}{\code{character}, field separator}

\item{dec}{\code{character}, decimal sign}
}
\value{
data.frame
}
\description{
Import timepoint data.
}
