% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importTimepoints}
\alias{importTimepoints}
\title{Import timepoint data.}
\usage{
importTimepoints(file)
}
\arguments{
\item{file}{\code{character}, filename}
}
\value{
data.frame
}
\description{
Import timepoint data.
}
