% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{sofaForTimepoints}
\alias{sofaForTimepoints}
\title{Get SOFA for specific timepoint}
\usage{
sofaForTimepoints(icm, tp, lag = 0L)
}
\arguments{
\item{icm}{\code{data.frame}, ICM data}

\item{tp}{\code{data.frme}, timepoint data, first column Id, other columns
timepoints}

\item{lag}{\code{numeric}, lag seconds added to reference date and extend the
range to 24 h + lag seconds (e.g. laboratory values take some time)}
}
\value{
\code{data.frame}
}
\description{
Get SOFA for specific timepoint
}
